requirements.txt
Scrapy>=2.8.0
scrapy-user-agents>=0.1.1
itemadapter>=0.3.0
Twisted>=22.10.0
markdownify>=0.11.6
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0